# fix: respect SIGTERM, faster Flagsmith API shutdowns